test(engine): rehome djot note coverage
Move the two Djot note bibliography checks from i18n into the Document - Djot Adapter suite so the behavioral report groups them correctly.

The pre-commit hook auto-bumped the workspace version to 0.26.1 as part of publishable crate bookkeeping.

Version-Bump: patch
