+++ Cargo.toml
@@ -17,7 +17,7 @@ members = [
 resolver = "2"
 
 [workspace.package]
-version = "0.26.0"
+version = "0.26.1"
 edition = "2024"
 license = "MIT OR Apache-2.0"
